Implementation of functions
set_grid_points_1D :
This function is a linespace from 0 to 1 with "la" steps which modifies teh vector X[..]

set_analytical_solution_DBC_1D :
This function applies the linesapce on BC0 and BC1 values by modifying EX_SOL value.

set_dense_RHS_DBS_1D :
This function modifies the first and last value of RHS and initializes the rest to 0.

set_GB_operator_colMajor_poisson1D_Id :
This function fills the diagonal of AB with 1.0 and initializes the rest to 0.

set_GB_operator_colMajor_poisson1D_Id :
This function modifies a compact band matrix (AB)
The first element of AB contains 0.0
The first line excepts the 1st elem is filled with -1.0
The second line with -2.0
The rest of the matrix except AB[i-1] for i={0,mat_size-1} filled with -1.0